refactor(telemetry): move settlement tests into a dedicated file
Review on #1679 asked for the tests in their own file rather than
an inline module. settlement.rs now declares `mod tests;` and the
module body lives unchanged in settlement/tests.rs, the same file
layout as agglayer-certificate-orchestrator's network_task tests.
